chore: move test setup to a helper function
This has no impact now but should make it easier to write benchmarks.
